Crate google-mapsengine1

This documentation was generated from Maps Engine crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

assets

get, list, parents.list and permissions.list

layers

cancel processing, create, delete, get, get published, list, list published, parents.list, patch, permissions.batch delete, permissions.batch update, permissions.list, process, publish and unpublish

maps

create, delete, get, get published, list, list published, patch, permissions.batch delete, permissions.batch update, permissions.list, publish and unpublish

projects

icons.create, icons.get, icons.list and list

raster collections

cancel processing, create, delete, get, list, parents.list, patch, permissions.batch delete, permissions.batch update, permissions.list, process, rasters.batch delete, rasters.batch insert and rasters.list

rasters

delete, files.insert, get, list, parents.list, patch, permissions.batch delete, permissions.batch update, permissions.list, process and upload

tables

create, delete, features.batch delete, features.batch insert, features.batch patch, features.get, features.list, files.insert, get, list, parents.list, patch, permissions.batch delete, permissions.batch update, permissions.list, process and upload

Upload supported by ...

icons create projects

files insert tables

files insert rasters

Download supported by ...

icons get projects

Everything else about the Maps Engine v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.tables().features_batch_insert(...).doit()
let r = hub.tables().list(...).doit()
let r = hub.tables().patch(...).doit()
let r = hub.tables().get(...).doit()
let r = hub.tables().files_insert(...).doit()
let r = hub.tables().features_batch_delete(...).doit()
let r = hub.tables().permissions_batch_delete(...).doit()
let r = hub.tables().permissions_batch_update(...).doit()
let r = hub.tables().features_list(...).doit()
let r = hub.tables().process(...).doit()
let r = hub.tables().upload(...).doit()
let r = hub.tables().delete(...).doit()
let r = hub.tables().parents_list(...).doit()
let r = hub.tables().features_batch_patch(...).doit()
let r = hub.tables().permissions_list(...).doit()
let r = hub.tables().features_get(...).doit()
let r = hub.tables().create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-mapsengine1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-mapsengine1" as mapsengine1;
use mapsengine1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tables().list()
             .tags("sea")
             .search("nonumy")
             .role("dolores")
             .project_id("gubergren")
             .processing_status("sadipscing")
             .page_token("aliquyam")
             .modified_before("ea")
             .modified_after("no")
             .max_results(80)
             .creator_email("justo")
             .created_before("et")
             .created_after("et")
             .bbox("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AcquisitionTime
Acquisition time represents acquired time of a raster.

Asset
An asset is any Google Maps Engine resource that has a globally unique ID. Assets include maps, layers, vector tables, raster collections, and rasters. Projects and features are not considered assets.

AssetGetCallBuilder
Return metadata for a particular asset.

AssetListCallBuilder
Return all assets readable by the current user.

AssetMethodsBuilder
A builder providing access to all methods supported on asset resources. It is not used directly, but through the MapsEngine hub.

AssetParentListCallBuilder
Return all parent ids of the specified asset.

AssetPermissionListCallBuilder
Return all of the permissions for the specified asset.

AssetsListResponse
The response returned by a call to resources.List.

Border
Border in line style. Both color and width are required.

Color
Basic color used in styling.

Datasource
There is no detailed description.

Datasources
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DisplayRule
A display rule of the vector style.

Feature
A feature within a table.

FeatureInfo
A feature info contains information about individual feature.

FeaturesBatchDeleteRequest
The request sent to features.BatchDelete.

FeaturesBatchInsertRequest
The request sent to features.Insert.

FeaturesBatchPatchRequest
The request sent to features.BatchPatch.

FeaturesListResponse
The response returned by a call to features.List.

File
A single File, which is a component of an Asset.

Filter
Conditions for filtering features.

GeoJsonGeometryCollection
A heterogenous collection of GeoJsonGeometry objects.

GeoJsonLineString
There is no detailed description.

GeoJsonMultiLineString
Multi Line String

GeoJsonMultiPoint
There is no detailed description.

GeoJsonMultiPolygon
There is no detailed description.

GeoJsonPoint
There is no detailed description.

GeoJsonPolygon
There is no detailed description.

GeoJsonPosition
A position represents a geographical position as an array containing a longitude and a latitude, and optionally an altitude, in that order. All Geometry objects make use of positions to represent geometries as nested arrays. The structure of the array is governed by the type of the geometry.

GeoJsonProperties
The properties associated with a feature.

Icon
An icon is a user-uploaded image that can be used to style point geometries.

IconStyle
Style for icon, this is part of point style.

IconsListResponse
The response returned by a call to icons.List.

LabelStyle
Text label style.

LatLngBox
A rectangular geographic bounds.

Layer
A Layer combines multiple datasources, with styling information, for presentation on a map.

LayerCancelProcessingCallBuilder
Cancel processing on a layer asset.

LayerCreateCallBuilder
Create a layer asset.

LayerDeleteCallBuilder
Delete a layer.

LayerGetCallBuilder
Return metadata for a particular layer.

LayerGetPublishedCallBuilder
Return the published metadata for a particular layer.

LayerListCallBuilder
Return all layers readable by the current user.

LayerListPublishedCallBuilder
Return all published layers readable by the current user.

LayerMethodsBuilder
A builder providing access to all methods supported on layer resources. It is not used directly, but through the MapsEngine hub.

LayerParentListCallBuilder
Return all parent ids of the specified layer.

LayerPatchCallBuilder
Mutate a layer asset.

LayerPermissionBatchDeleteCallBuilder
Remove permission entries from an already existing asset.

LayerPermissionBatchUpdateCallBuilder
Add or update permission entries to an already existing asset.

LayerPermissionListCallBuilder
Return all of the permissions for the specified asset.

LayerProcesCallBuilder
Process a layer asset.

LayerPublishCallBuilder
Publish a layer asset.

LayerUnpublishCallBuilder
Unpublish a layer asset.

LayersListResponse
The response returned by a call to layers.List. Note: The list response does not include all the fields available in a layer. Refer to the layer resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each layer.

LineStyle
Style for lines.

LineStyleStroke
Stroke of the line.

Map
A Map is a collection of Layers, optionally contained within folders.

MapContents
There is no detailed description.

MapCreateCallBuilder
Create a map asset.

MapDeleteCallBuilder
Delete a map.

MapFolder
There is no detailed description.

MapGetCallBuilder
Return metadata for a particular map.

MapGetPublishedCallBuilder
Return the published metadata for a particular map.

MapKmlLink
There is no detailed description.

MapLayer
There is no detailed description.

MapListCallBuilder
Return all maps readable by the current user.

MapListPublishedCallBuilder
Return all published maps readable by the current user.

MapMethodsBuilder
A builder providing access to all methods supported on map resources. It is not used directly, but through the MapsEngine hub.

MapPatchCallBuilder
Mutate a map asset.

MapPermissionBatchDeleteCallBuilder
Remove permission entries from an already existing asset.

MapPermissionBatchUpdateCallBuilder
Add or update permission entries to an already existing asset.

MapPermissionListCallBuilder
Return all of the permissions for the specified asset.

MapPublishCallBuilder
Publish a map asset.

MapUnpublishCallBuilder
Unpublish a map asset.

MapsEngine
Central instance to access all MapsEngine related resource activities

MapsListResponse
The response returned by a call to maps.List.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Parent
A list of the parents of an asset.

ParentsListResponse
The response returned by a call to parents.List.

Permission
A permission defines the user or group that has access to an asset, and the type of access they have.

PermissionsBatchDeleteRequest
The request sent to mapsengine.permissions.batchDelete.

PermissionsBatchDeleteResponse
The response returned by a call to mapsengine.permissions.batchDelete.

PermissionsBatchUpdateRequest
The request sent to mapsengine.permissions.batchUpdate.

PermissionsBatchUpdateResponse
The response returned by a call to mapsengine.permissions.batchUpdate.

PermissionsListResponse
There is no detailed description.

PointStyle
Style for points.

PolygonStyle
Style for polygons.

ProcessResponse
The response returned by a call to any asset's Process method.

Project
A Maps Engine project groups a collection of resources.

ProjectIconCreateCallBuilder
Create an icon.

ProjectIconGetCallBuilder
Return an icon or its associated metadata

ProjectIconListCallBuilder
Return all icons in the current project

ProjectListCallBuilder
Return all projects readable by the current user.

ProjectMethodsBuilder
A builder providing access to all methods supported on project resources. It is not used directly, but through the MapsEngine hub.

ProjectsListResponse
The response returned by a call to projects.List.

PublishResponse
The response returned by a call to any asset's Publish method.

PublishedLayer
The published version of a layer.

PublishedLayersListResponse
The response returned by a call to layers.List.published.

PublishedMap
The published version of a map asset.

PublishedMapsListResponse
The response returned by a call to maps.List.published.

Raster
A geo-referenced raster.

RasterCollection
A raster collection groups multiple Raster resources for inclusion in a Layer.

RasterCollectionCancelProcessingCallBuilder
Cancel processing on a raster collection asset.

RasterCollectionCreateCallBuilder
Create a raster collection asset.

RasterCollectionDeleteCallBuilder
Delete a raster collection.

RasterCollectionGetCallBuilder
Return metadata for a particular raster collection.

RasterCollectionListCallBuilder
Return all raster collections readable by the current user.

RasterCollectionMethodsBuilder
A builder providing access to all methods supported on rasterCollection resources. It is not used directly, but through the MapsEngine hub.

RasterCollectionParentListCallBuilder
Return all parent ids of the specified raster collection.

RasterCollectionPatchCallBuilder
Mutate a raster collection asset.

RasterCollectionPermissionBatchDeleteCallBuilder
Remove permission entries from an already existing asset.

RasterCollectionPermissionBatchUpdateCallBuilder
Add or update permission entries to an already existing asset.

RasterCollectionPermissionListCallBuilder
Return all of the permissions for the specified asset.

RasterCollectionProcesCallBuilder
Process a raster collection asset.

RasterCollectionRasterBatchDeleteCallBuilder
Remove rasters from an existing raster collection.

RasterCollectionRasterBatchInsertCallBuilder
Add rasters to an existing raster collection. Rasters must be successfully processed in order to be added to a raster collection.

RasterCollectionRasterListCallBuilder
Return all rasters within a raster collection.

RasterCollectionsListResponse
The response returned by a call to raster_collections.List. Note: The list response does not include all the fields available in a raster collection. Refer to the RasterCollection resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each raster collection.

RasterCollectionsRaster
A raster resource.

RasterCollectionsRasterBatchDeleteRequest
The request sent to rasterCollections.Rasters.BatchDelete.

RasterCollectionsRastersBatchDeleteResponse
The response returned by a call to rasterCollections.rasters.batchDelete.

RasterCollectionsRastersBatchInsertRequest
The request sent to rasterCollections.Rasters.BatchInsert.

RasterCollectionsRastersBatchInsertResponse
The response returned by a call to rasterCollections.rasters.batchInsert.

RasterCollectionsRastersListResponse
The response returned by a call to rasterCollections.rasters.List.

RasterDeleteCallBuilder
Delete a raster.

RasterFileInsertCallBuilder
Upload a file to a raster asset.

RasterGetCallBuilder
Return metadata for a single raster.

RasterListCallBuilder
Return all rasters readable by the current user.

RasterMethodsBuilder
A builder providing access to all methods supported on raster resources. It is not used directly, but through the MapsEngine hub.

RasterParentListCallBuilder
Return all parent ids of the specified rasters.

RasterPatchCallBuilder
Mutate a raster asset.

RasterPermissionBatchDeleteCallBuilder
Remove permission entries from an already existing asset.

RasterPermissionBatchUpdateCallBuilder
Add or update permission entries to an already existing asset.

RasterPermissionListCallBuilder
Return all of the permissions for the specified asset.

RasterProcesCallBuilder
Process a raster asset.

RasterUploadCallBuilder
Create a skeleton raster asset for upload.

RastersListResponse
The response returned by a call to rasters.List.

ScaledShape
Parameters for styling points as scaled shapes.

ScalingFunction
Parameters for scaling scaled shapes.

Schema
A schema indicating the properties which may be associated with features within a Table, and the types of those properties.

SizeRange
Scaled shape size range in pixels. For circles, size corresponds to diameter.

Table
A collection of geographic features, and associated metadata.

TableColumn
There is no detailed description.

TableCreateCallBuilder
Create a table asset.

TableDeleteCallBuilder
Delete a table.

TableFeatureBatchDeleteCallBuilder
Delete all features matching the given IDs.

TableFeatureBatchInsertCallBuilder
Append features to an existing table.

TableFeatureBatchPatchCallBuilder
Update the supplied features.

TableFeatureGetCallBuilder
Return a single feature, given its ID.

TableFeatureListCallBuilder
Return all features readable by the current user.

TableFileInsertCallBuilder
Upload a file to a placeholder table asset. See Table Upload in the Developer's Guide for more information. Supported file types are listed in the Supported data formats and limits article of the Google Maps Engine help center.

TableGetCallBuilder
Return metadata for a particular table, including the schema.

TableListCallBuilder
Return all tables readable by the current user.

TableMethodsBuilder
A builder providing access to all methods supported on table resources. It is not used directly, but through the MapsEngine hub.

TableParentListCallBuilder
Return all parent ids of the specified table.

TablePatchCallBuilder
Mutate a table asset.

TablePermissionBatchDeleteCallBuilder
Remove permission entries from an already existing asset.

TablePermissionBatchUpdateCallBuilder
Add or update permission entries to an already existing asset.

TablePermissionListCallBuilder
Return all of the permissions for the specified asset.

TableProcesCallBuilder
Process a table asset.

TableUploadCallBuilder
Create a placeholder table asset to which table files can be uploaded. Once the placeholder has been created, files are uploaded to the https://www.googleapis.com/upload/mapsengine/v1/tables/table_id/files endpoint. See Table Upload in the Developer's Guide or Table.files: insert in the reference documentation for more information.

TablesListResponse
The response returned by a call to tables.List. Note: The list response does not include all the fields available in a table. Refer to the table resource description for details of the fields that are not included. You'll need to send a get request to retrieve the additional fields for each table.

Tags
There is no detailed description.

ValueRange
Range of values used for scaling shapes. The min/max values will be drawn as shapes with the min/max size.

VectorStyle
A vector style contains styling information for vector layer.

ZoomLevels
Zoom level range. Zoom levels are restricted between 0 and 24, inclusive.

Enums

GeoJsonGeometry
There is no detailed description.

MapItem
There is no detailed description.

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.