Struct google-mapsengine1::LayerMethodsBuilder

pub struct LayerMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on layer resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-mapsengine1" as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `cancel_processing(...)`, `create(...)`, `delete(...)`, `get(...)`, `get_published(...)`, `list(...)`, `list_published(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)`, `publish(...)` and `unpublish(...)`
// to build up your call.
let rb = hub.layers();

Methods

impl<'a, C, NC, A> LayerMethodsBuilder<'a, C, NC, A>

fn unpublish(&self, id: &str) -> LayerUnpublishMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Unpublish a layer asset.

fn publish(&self, id: &str) -> LayerPublishMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Publish a layer asset.

fn list_published(&self) -> LayerListPublishedMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all published layers readable by the current user.

fn create(&self, request: &Layer) -> LayerCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a layer asset.

fn get_published(&self, id: &str) -> LayerGetPublishedMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return the published metadata for a particular layer.

fn delete(&self, id: &str) -> LayerDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete a layer.

fn get(&self, id: &str) -> LayerGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return metadata for a particular layer.

fn permissions_batch_delete(&self, request: &PermissionsBatchDeleteRequest, id: &str) -> LayerPermissionBatchDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

fn permissions_batch_update(&self, request: &PermissionsBatchUpdateRequest, id: &str) -> LayerPermissionBatchUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

fn list(&self) -> LayerListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all layers readable by the current user.

fn parents_list(&self, id: &str) -> LayerParentListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified layer.

fn cancel_processing(&self, id: &str) -> LayerCancelProcessingMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Cancel processing on a layer asset.

fn patch(&self, request: &Layer, id: &str) -> LayerPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Mutate a layer asset.

fn permissions_list(&self, id: &str) -> LayerPermissionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

fn process(&self, id: &str) -> LayerProcesMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Process a layer asset.