Struct google-mapsengine1::Map

pub struct Map {
    pub description: Option<String>,
    pub default_viewport: Option<LatLngBox>,
    pub tags: Option<Tags>,
    pub draft_access_list: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub creator_email: Option<String>,
    pub bbox: Vec<f64>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub contents: Option<MapContents>,
    pub name: Option<String>,
    pub processing_status: Option<String>,
    pub project_id: Option<String>,
    pub versions: Vec<String>,
    pub creation_time: Option<String>,
    pub publishing_status: Option<String>,
    pub etag: Option<String>,
    pub published_access_list: Option<String>,
}

A Map is a collection of Layers, optionally contained within folders.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get published maps (none)

get maps (response)

create maps (request|response)

list published maps (none)

list maps (none)

publish maps (none)

patch maps (request)

permissions list maps (none)

permissions batch delete maps (none)

unpublish maps (none)

permissions batch update maps (none)

delete maps (none)

Fields

description

The description of this Map, supplied by the author.

default_viewport

An array of four numbers (west, south, east, north) which defines the rectangular bounding box of the default viewport. The numbers represent latitude and longitude in decimal degrees.

tags

Tags of this Map.

draft_access_list

Deprecated: The name of an access list of the Map Editor type. The user on whose behalf the request is being sent must be an editor on that access list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

creator_email

The email address of the creator of this map. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this Map. The box is expressed as \"west, south, east, north\". The numbers represent latitude and longitude in decimal degrees.

last_modifier_email

The email address of the last modifier of this map. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this map. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this Map.

contents

The contents of this Map.

name

The name of this Map, supplied by the author.

processing_status

The processing status of this map. Map processing is automatically started once a map becomes ready for processing.

project_id

The ID of the project that this Map is in.

versions

Deprecated: An array containing the available versions of this Map. Currently may only contain "published". The publishingStatus field should be used instead.

creation_time

The creation time of this map. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

publishing_status

The publishing status of this map.

etag

The ETag, used to refer to the current version of the asset.

published_access_list

Deprecated: The access list to whom view permissions are granted. The value must be the name of a Maps Engine access list of the Map Viewer type, and the user must be a viewer on that list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. This is an input field only. It is not returned in response to a list or get request.