Struct google-mapsengine1::MapMethods

pub struct MapMethods<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on map resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "google-mapsengine1" as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `get_published(...)`, `list(...)`, `list_published(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `publish(...)` and `unpublish(...)`
// to build up your call.
let rb = hub.maps();

Methods

impl<'a, C, NC, A> MapMethods<'a, C, NC, A>

fn list(&self) -> MapListCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all maps readable by the current user.

fn list_published(&self) -> MapListPublishedCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all published maps readable by the current user.

fn patch(&self, request: &Map, id: &str) -> MapPatchCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Mutate a map asset.

fn permissions_batch_update(&self, request: &PermissionsBatchUpdateRequest, id: &str) -> MapPermissionBatchUpdateCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

fn delete(&self, id: &str) -> MapDeleteCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete a map.

fn unpublish(&self, id: &str) -> MapUnpublishCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Unpublish a map asset.

fn publish(&self, id: &str) -> MapPublishCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Publish a map asset.

fn permissions_list(&self, id: &str) -> MapPermissionListCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

fn create(&self, request: &Map) -> MapCreateCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a map asset.

fn get_published(&self, id: &str) -> MapGetPublishedCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Return the published metadata for a particular map.

fn get(&self, id: &str) -> MapGetCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Return metadata for a particular map.

fn permissions_batch_delete(&self, request: &PermissionsBatchDeleteRequest, id: &str) -> MapPermissionBatchDeleteCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.