Struct google_mapsengine1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `icons_create(...)`, `icons_get(...)`, `icons_list(...)` and `list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn icons_get(&self, project_id: &str, id: &str) -> ProjectIconGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return an icon or its associated metadata

Arguments

projectId - The ID of the project.

id - The ID of the icon.

fn icons_create(&self, request: Icon, project_id: &str) -> ProjectIconCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create an icon.

Arguments

request - No description provided.

projectId - The ID of the project.

fn list(&self) -> ProjectListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all projects readable by the current user.

fn icons_list(&self, project_id: &str) -> ProjectIconListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all icons in the current project

Arguments

projectId - The ID of the project.