Struct google_mapsengine1::Raster

pub struct Raster {
    pub files: Option<Vec<File>>,
    pub attribution: Option<String>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub draft_access_list: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub creator_email: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
    pub processing_status: Option<String>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub etag: Option<String>,
    pub acquisition_time: Option<AcquisitionTime>,
    pub mask_type: Option<String>,
    pub raster_type: Option<String>,
}

A geo-referenced raster.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

permissions batch update rasters (none)

patch rasters (request)

process rasters (none)

permissions list rasters (none)

delete rasters (none)

get rasters (response)

list rasters (none)

permissions batch delete rasters (none)

parents list rasters (none)

upload rasters (request|response)

files insert rasters (none)

Fields

files

The files associated with this Raster.

attribution

The name of the attribution to be used for this Raster.

description

The description of this Raster, supplied by the author.

tags

Tags of this Raster.

draft_access_list

Deprecated: The name of an access list of the Map Editor type. The user on whose behalf the request is being sent must be an editor on that access list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

creator_email

The email address of the creator of this raster. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this Raster. The box is expressed as \"west, south, east, north\". The numbers represent latitudes and longitudes in decimal degrees.

last_modifier_email

The email address of the last modifier of this raster. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this raster. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this Raster.

name

The name of this Raster, supplied by the author.

processing_status

The processing status of this Raster.

project_id

The ID of the project that this Raster is in.

creation_time

The creation time of this raster. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

etag

The ETag, used to refer to the current version of the asset.

acquisition_time

The acquisition time of this Raster.

mask_type

The mask processing type of this Raster.

raster_type

The type of this Raster. Always "image" today.