Struct google_mapsengine1::RasterCollection

pub struct RasterCollection {
    pub attribution: Option<String>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub draft_access_list: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub creator_email: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub mosaic: Option<bool>,
    pub name: Option<String>,
    pub processing_status: Option<String>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub etag: Option<String>,
    pub raster_type: Option<String>,
}

A raster collection groups multiple Raster resources for inclusion in a Layer.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

create raster collections (request|response)

parents list raster collections (none)

permissions list raster collections (none)

list raster collections (none)

process raster collections (none)

permissions batch update raster collections (none)

get raster collections (response)

permissions batch delete raster collections (none)

patch raster collections (request)

cancel processing raster collections (none)

rasters batch insert raster collections (none)

rasters batch delete raster collections (none)

delete raster collections (none)

rasters list raster collections (none)

Fields

attribution

The name of the attribution to be used for this RasterCollection. Note: Attribution is returned in response to a get request but not a list request. After requesting a list of raster collections, you'll need to send a get request to retrieve the attribution for each raster collection.

description

The description of this RasterCollection, supplied by the author.

tags

Tags of this RasterCollection.

draft_access_list

Deprecated: The name of an access list of the Map Editor type. The user on whose behalf the request is being sent must be an editor on that access list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

creator_email

The email address of the creator of this raster collection. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this RasterCollection. The box is expressed as \"west, south, east, north\". The numbers represent latitude and longitude in decimal degrees.

last_modifier_email

The email address of the last modifier of this raster collection. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this RasterCollection. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this RasterCollection.

mosaic

True if this RasterCollection is a mosaic.

name

The name of this RasterCollection, supplied by the author.

processing_status

The processing status of this RasterCollection.

project_id

The ID of the project that this RasterCollection is in.

creation_time

The creation time of this RasterCollection. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

etag

The ETag, used to refer to the current version of the asset.

raster_type

The type of rasters contained within this RasterCollection.