Struct google_mapsengine1::RasterCollectionMethods

pub struct RasterCollectionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on rasterCollection resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `cancel_processing(...)`, `create(...)`, `delete(...)`, `get(...)`, `list(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)`, `rasters_batch_delete(...)`, `rasters_batch_insert(...)` and `rasters_list(...)`
// to build up your call.
let rb = hub.raster_collections();

Methods

impl<'a, C, A> RasterCollectionMethods<'a, C, A>

fn create(&self, request: RasterCollection) -> RasterCollectionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create a raster collection asset.

Arguments

request - No description provided.

fn permissions_batch_delete(&self, request: PermissionsBatchDeleteRequest, id: &str) -> RasterCollectionPermissionBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

Arguments

request - No description provided.

id - The ID of the asset from which permissions will be removed.

fn patch(&self, request: RasterCollection, id: &str) -> RasterCollectionPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Mutate a raster collection asset.

Arguments

request - No description provided.

id - The ID of the raster collection.

fn cancel_processing(&self, id: &str) -> RasterCollectionCancelProcessingCall<'a, C, A>

Create a builder to help you perform the following task:

Cancel processing on a raster collection asset.

Arguments

id - The ID of the raster collection.

fn process(&self, id: &str) -> RasterCollectionProcesCall<'a, C, A>

Create a builder to help you perform the following task:

Process a raster collection asset.

Arguments

id - The ID of the raster collection.

fn permissions_batch_update(&self, request: PermissionsBatchUpdateRequest, id: &str) -> RasterCollectionPermissionBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the asset to which permissions will be added.

fn delete(&self, id: &str) -> RasterCollectionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a raster collection.

Arguments

id - The ID of the raster collection. Only the raster collection creator or project owner are permitted to delete. If the rastor collection is included in a layer, the request will fail. Remove the raster collection from all layers prior to deleting.

fn parents_list(&self, id: &str) -> RasterCollectionParentListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified raster collection.

Arguments

id - The ID of the raster collection whose parents will be listed.

fn list(&self) -> RasterCollectionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all raster collections readable by the current user.

fn rasters_batch_insert(&self, request: RasterCollectionsRastersBatchInsertRequest, id: &str) -> RasterCollectionRasterBatchInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Add rasters to an existing raster collection. Rasters must be successfully processed in order to be added to a raster collection.

Up to 50 rasters can be included in a single batchInsert request. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the raster collection to which these rasters belong.

fn rasters_batch_delete(&self, request: RasterCollectionsRasterBatchDeleteRequest, id: &str) -> RasterCollectionRasterBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove rasters from an existing raster collection.

Up to 50 rasters can be included in a single batchDelete request. Each batchDelete request is atomic.

Arguments

request - No description provided.

id - The ID of the raster collection to which these rasters belong.

fn permissions_list(&self, id: &str) -> RasterCollectionPermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

Arguments

id - The ID of the asset whose permissions will be listed.

fn get(&self, id: &str) -> RasterCollectionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return metadata for a particular raster collection.

Arguments

id - The ID of the raster collection.

fn rasters_list(&self, id: &str) -> RasterCollectionRasterListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all rasters within a raster collection.

Arguments

id - The ID of the raster collection to which these rasters belong.