Struct google_mapsengine1::RasterMethods

pub struct RasterMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on raster resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `files_insert(...)`, `get(...)`, `list(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)` and `upload(...)`
// to build up your call.
let rb = hub.rasters();

Methods

impl<'a, C, A> RasterMethods<'a, C, A>

fn permissions_batch_update(&self, request: PermissionsBatchUpdateRequest, id: &str) -> RasterPermissionBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the asset to which permissions will be added.

fn patch(&self, request: Raster, id: &str) -> RasterPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Mutate a raster asset.

Arguments

request - No description provided.

id - The ID of the raster.

fn permissions_list(&self, id: &str) -> RasterPermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

Arguments

id - The ID of the asset whose permissions will be listed.

fn delete(&self, id: &str) -> RasterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a raster.

Arguments

id - The ID of the raster. Only the raster creator or project owner are permitted to delete. If the raster is included in a layer or mosaic, the request will fail. Remove it from all parents prior to deleting.

fn permissions_batch_delete(&self, request: PermissionsBatchDeleteRequest, id: &str) -> RasterPermissionBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

Arguments

request - No description provided.

id - The ID of the asset from which permissions will be removed.

fn files_insert(&self, id: &str, filename: &str) -> RasterFileInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Upload a file to a raster asset.

Arguments

id - The ID of the raster asset.

filename - The file name of this uploaded file.

fn process(&self, id: &str) -> RasterProcesCall<'a, C, A>

Create a builder to help you perform the following task:

Process a raster asset.

Arguments

id - The ID of the raster.

fn get(&self, id: &str) -> RasterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return metadata for a single raster.

Arguments

id - The ID of the raster.

fn list(&self, project_id: &str) -> RasterListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all rasters readable by the current user.

Arguments

projectId - The ID of a Maps Engine project, used to filter the response. To list all available projects with their IDs, send a Projects: list request. You can also find your project ID as the value of the DashboardPlace:cid URL parameter when signed in to mapsengine.google.com.

fn upload(&self, request: Raster) -> RasterUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Create a skeleton raster asset for upload.

Arguments

request - No description provided.

fn parents_list(&self, id: &str) -> RasterParentListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified rasters.

Arguments

id - The ID of the rasters whose parents will be listed.