Struct google_mapsengine1::Table

pub struct Table {
    pub files: Option<Vec<File>>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub draft_access_list: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub creator_email: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
    pub processing_status: Option<String>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub source_encoding: Option<String>,
    pub etag: Option<String>,
    pub published_access_list: Option<String>,
    pub schema: Option<Schema>,
}

A collection of geographic features, and associated metadata.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

features batch insert tables (none)

list tables (none)

patch tables (request)

get tables (response)

files insert tables (none)

features batch delete tables (none)

permissions batch delete tables (none)

permissions batch update tables (none)

features list tables (none)

process tables (none)

upload tables (request|response)

delete tables (none)

parents list tables (none)

features batch patch tables (none)

permissions list tables (none)

features get tables (none)

create tables (request|response)

Fields

files

The files associated with this table.

description

The description of this table, supplied by the author.

tags

An array of text strings, with each string representing a tag. More information about tags can be found in the Tagging data article of the Maps Engine help center.

draft_access_list

Deprecated: The name of an access list of the Map Editor type. The user on whose behalf the request is being sent must be an editor on that access list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

creator_email

The email address of the creator of this table. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this Table. The box is expressed as \"west, south, east, north\". The numbers represent latitude and longitude in decimal degrees.

last_modifier_email

The email address of the last modifier of this table. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this table. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this table.

name

The name of this table, supplied by the author.

processing_status

The processing status of this table.

project_id

The ID of the project to which the table belongs.

creation_time

The creation time of this table. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

source_encoding

Encoding of the uploaded files. Valid values include UTF-8, CP1251, ISO 8859-1, and Shift_JIS.

etag

The ETag, used to refer to the current version of the asset.

published_access_list

Deprecated: The access list to whom view permissions are granted. The value must be the name of a Maps Engine access list of the Map Viewer type, and the user must be a viewer on that list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

schema

The schema for this table. Note: The schema is returned in response to a get request but not a list request. After requesting a list of tables, you'll need to send a get request to retrieve the schema for each table.