Struct google_mapsengine1::TableListCall

pub struct TableListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Return all tables readable by the current user.

A builder for the list method supported by a table resource. It is not used directly, but through a TableMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tables().list()
             .tags("labore")
             .search("sed")
             .role("ea")
             .project_id("gubergren")
             .processing_status("aliquyam")
             .page_token("eos")
             .modified_before("tempor")
             .modified_after("sea")
             .max_results(16)
             .creator_email("ipsum")
             .created_before("aliquyam")
             .created_after("dolores")
             .bbox("sit")
             .doit();

Methods

impl<'a, C, A> TableListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, TablesListResponse)>

Perform the operation you have build so far.

fn tags(self, new_value: &str) -> TableListCall<'a, C, A>

A comma separated list of tags. Returned assets will contain all the tags from the list.

Sets the tags query property to the given value.

fn search(self, new_value: &str) -> TableListCall<'a, C, A>

An unstructured search string used to filter the set of results based on asset metadata.

Sets the search query property to the given value.

fn role(self, new_value: &str) -> TableListCall<'a, C, A>

The role parameter indicates that the response should only contain assets where the current user has the specified level of access.

Sets the role query property to the given value.

fn project_id(self, new_value: &str) -> TableListCall<'a, C, A>

The ID of a Maps Engine project, used to filter the response. To list all available projects with their IDs, send a Projects: list request. You can also find your project ID as the value of the DashboardPlace:cid URL parameter when signed in to mapsengine.google.com.

Sets the project id query property to the given value.

fn processing_status(self, new_value: &str) -> TableListCall<'a, C, A>

Sets the processing status query property to the given value.

fn page_token(self, new_value: &str) -> TableListCall<'a, C, A>

The continuation token, used to page through large result sets. To get the next page of results, set this parameter to the value of nextPageToken from the previous response.

Sets the page token query property to the given value.

fn modified_before(self, new_value: &str) -> TableListCall<'a, C, A>

An RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z). Returned assets will have been modified at or before this time.

Sets the modified before query property to the given value.

fn modified_after(self, new_value: &str) -> TableListCall<'a, C, A>

An RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z). Returned assets will have been modified at or after this time.

Sets the modified after query property to the given value.

fn max_results(self, new_value: u32) -> TableListCall<'a, C, A>

The maximum number of items to include in a single response page. The maximum supported value is 100.

Sets the max results query property to the given value.

fn creator_email(self, new_value: &str) -> TableListCall<'a, C, A>

An email address representing a user. Returned assets that have been created by the user associated with the provided email address.

Sets the creator email query property to the given value.

fn created_before(self, new_value: &str) -> TableListCall<'a, C, A>

An RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z). Returned assets will have been created at or before this time.

Sets the created before query property to the given value.

fn created_after(self, new_value: &str) -> TableListCall<'a, C, A>

An RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z). Returned assets will have been created at or after this time.

Sets the created after query property to the given value.

fn bbox(self, new_value: &str) -> TableListCall<'a, C, A>

A bounding box, expressed as "west,south,east,north". If set, only assets which intersect this bounding box will be returned.

Sets the bbox query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> TableListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> TableListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> TableListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.