Struct google_mapsengine1::TableMethods

pub struct TableMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on table resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `features_batch_delete(...)`, `features_batch_insert(...)`, `features_batch_patch(...)`, `features_get(...)`, `features_list(...)`, `files_insert(...)`, `get(...)`, `list(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)` and `upload(...)`
// to build up your call.
let rb = hub.tables();

Methods

impl<'a, C, A> TableMethods<'a, C, A>

fn delete(&self, id: &str) -> TableDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a table.

Arguments

id - The ID of the table. Only the table creator or project owner are permitted to delete. If the table is included in a layer, the request will fail. Remove it from all layers prior to deleting.

fn files_insert(&self, id: &str, filename: &str) -> TableFileInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Upload a file to a placeholder table asset. See Table Upload in the Developer's Guide for more information. Supported file types are listed in the Supported data formats and limits article of the Google Maps Engine help center.

Arguments

id - The ID of the table asset.

filename - The file name of this uploaded file.

fn features_batch_insert(&self, request: FeaturesBatchInsertRequest, id: &str) -> TableFeatureBatchInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Append features to an existing table.

A single batchInsert request can create:

Up to 50 features.

A combined total of 10 000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

For more information about inserting features, read Creating features in the Google Maps Engine developer's guide.

Arguments

request - No description provided.

id - The ID of the table to append the features to.

fn list(&self) -> TableListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all tables readable by the current user.

fn get(&self, id: &str) -> TableGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return metadata for a particular table, including the schema.

Arguments

id - The ID of the table.

fn features_batch_delete(&self, request: FeaturesBatchDeleteRequest, id: &str) -> TableFeatureBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete all features matching the given IDs.

Arguments

request - No description provided.

id - The ID of the table that contains the features to be deleted.

fn features_list(&self, id: &str) -> TableFeatureListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all features readable by the current user.

Arguments

id - The ID of the table to which these features belong.

fn process(&self, id: &str) -> TableProcesCall<'a, C, A>

Create a builder to help you perform the following task:

Process a table asset.

Arguments

id - The ID of the table.

fn parents_list(&self, id: &str) -> TableParentListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified table.

Arguments

id - The ID of the table whose parents will be listed.

fn permissions_batch_update(&self, request: PermissionsBatchUpdateRequest, id: &str) -> TablePermissionBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the asset to which permissions will be added.

fn upload(&self, request: Table) -> TableUploadCall<'a, C, A>

Create a builder to help you perform the following task:

Create a placeholder table asset to which table files can be uploaded. Once the placeholder has been created, files are uploaded to the https://www.googleapis.com/upload/mapsengine/v1/tables/table_id/files endpoint. See Table Upload in the Developer's Guide or Table.files: insert in the reference documentation for more information.

Arguments

request - No description provided.

fn features_batch_patch(&self, request: FeaturesBatchPatchRequest, id: &str) -> TableFeatureBatchPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Update the supplied features.

A single batchPatch request can update:

Up to 50 features.

A combined total of 10 000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

Feature updates use HTTP PATCH semantics:

A supplied value replaces an existing value (if any) in that field.

Omitted fields remain unchanged.

Complex values in geometries and properties must be replaced as atomic units. For example, providing just the coordinates of a geometry is not allowed; the complete geometry, including type, must be supplied.

Setting a property's value to null deletes that property. For more information about updating features, read Updating features in the Google Maps Engine developer's guide.

Arguments

request - No description provided.

id - The ID of the table containing the features to be patched.

fn patch(&self, request: Table, id: &str) -> TablePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Mutate a table asset.

Arguments

request - No description provided.

id - The ID of the table.

fn permissions_batch_delete(&self, request: PermissionsBatchDeleteRequest, id: &str) -> TablePermissionBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

Arguments

request - No description provided.

id - The ID of the asset from which permissions will be removed.

fn create(&self, request: Table) -> TableCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create a table asset.

Arguments

request - No description provided.

fn permissions_list(&self, id: &str) -> TablePermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

Arguments

id - The ID of the asset whose permissions will be listed.

fn features_get(&self, table_id: &str, id: &str) -> TableFeatureGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return a single feature, given its ID.

Arguments

tableId - The ID of the table.

id - The ID of the feature to get.