Struct google-mapsengine1::TableMethodsBuilder

pub struct TableMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on table resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-mapsengine1" as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `create(...)`, `delete(...)`, `features_batch_delete(...)`, `features_batch_insert(...)`, `features_batch_patch(...)`, `features_get(...)`, `features_list(...)`, `files_insert(...)`, `get(...)`, `list(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)` and `upload(...)`
// to build up your call.
let rb = hub.tables();

Methods

impl<'a, C, NC, A> TableMethodsBuilder<'a, C, NC, A>

fn delete(&self, id: &str) -> TableDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete a table.

fn files_insert(&self, id: &str, filename: &str) -> TableFileInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Upload a file to a placeholder table asset. See Table Upload in the Developer's Guide for more information. Supported file types are listed in the Supported data formats and limits article of the Google Maps Engine help center.

fn features_batch_insert(&self, request: &FeaturesBatchInsertRequest, id: &str) -> TableFeatureBatchInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Append features to an existing table.

A single batchInsert request can create:

Up to 50 features.

A combined total of 10 000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

For more information about inserting features, read Creating features in the Google Maps Engine developer's guide.

fn list(&self) -> TableListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all tables readable by the current user.

fn get(&self, id: &str) -> TableGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return metadata for a particular table, including the schema.

fn features_batch_delete(&self, request: &FeaturesBatchDeleteRequest, id: &str) -> TableFeatureBatchDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Delete all features matching the given IDs.

fn features_list(&self, id: &str) -> TableFeatureListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all features readable by the current user.

fn process(&self, id: &str) -> TableProcesMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Process a table asset.

fn parents_list(&self, id: &str) -> TableParentListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified table.

fn permissions_batch_update(&self, request: &PermissionsBatchUpdateRequest, id: &str) -> TablePermissionBatchUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

fn upload(&self, request: &Table) -> TableUploadMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a placeholder table asset to which table files can be uploaded. Once the placeholder has been created, files are uploaded to the https://www.googleapis.com/upload/mapsengine/v1/tables/table_id/files endpoint. See Table Upload in the Developer's Guide or Table.files: insert in the reference documentation for more information.

fn features_batch_patch(&self, request: &FeaturesBatchPatchRequest, id: &str) -> TableFeatureBatchPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Update the supplied features.

A single batchPatch request can update:

Up to 50 features.

A combined total of 10 000 vertices. Feature limits are documented in the Supported data formats and limits article of the Google Maps Engine help center. Note that free and paid accounts have different limits.

Feature updates use HTTP PATCH semantics:

A supplied value replaces an existing value (if any) in that field.

Omitted fields remain unchanged.

Complex values in geometries and properties must be replaced as atomic units. For example, providing just the coordinates of a geometry is not allowed; the complete geometry, including type, must be supplied.

Setting a property's value to null deletes that property. For more information about updating features, read Updating features in the Google Maps Engine developer's guide.

fn patch(&self, request: &Table, id: &str) -> TablePatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Mutate a table asset.

fn permissions_batch_delete(&self, request: &PermissionsBatchDeleteRequest, id: &str) -> TablePermissionBatchDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

fn create(&self, request: &Table) -> TableCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a table asset.

fn permissions_list(&self, id: &str) -> TablePermissionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

fn features_get(&self, table_id: &str, id: &str) -> TableFeatureGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Return a single feature, given its ID.