Struct google_mapsengine1::Asset

pub struct Asset {
    pub resource: Option<String>,
    pub description: Option<String>,
    pub tags: Option<Vec<String>>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub etag: Option<String>,
    pub creator_email: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
}

An asset is any Google Maps Engine resource that has a globally unique ID. Assets include maps, layers, vector tables, raster collections, and rasters. Projects and features are not considered assets.

More detailed information about an asset can be obtained by querying the asset's particular endpoint.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

parents list assets (none)

list assets (none)

permissions list assets (none)

get assets (response)

Fields

resource

The URL to query to retrieve the asset's complete object. The assets endpoint only returns high-level information about a resource.

description

The asset's description.

tags

An array of text strings, with each string representing a tag. More information about tags can be found in the Tagging data article of the Maps Engine help center.

project_id

The ID of the project to which the asset belongs.

creation_time

The creation time of this asset. The value is an RFC 3339-formatted date-time value (for example, 1970-01-01T00:00:00Z).

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

etag

The ETag, used to refer to the current version of the asset.

creator_email

The email address of the creator of this asset. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this asset. The box is expressed as "west, south, east, north". The numbers represent latitude and longitude in decimal degrees.

last_modifier_email

The email address of the last modifier of this asset. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this asset. The value is an RFC 3339-formatted date-time value (for example, 1970-01-01T00:00:00Z).

type_

The type of asset. One of raster, rasterCollection, table, map, or layer.

id

The asset's globally unique ID.

name

The asset's name.