Struct google_mapsengine1::Layer

pub struct Layer {
    pub datasources: Option<Datasources>,
    pub tags: Option<Tags>,
    pub draft_access_list: Option<String>,
    pub datasource_type: Option<String>,
    pub writers_can_edit_permissions: Option<bool>,
    pub creator_email: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modifier_email: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub description: Option<String>,
    pub style: Option<VectorStyle>,
    pub name: Option<String>,
    pub processing_status: Option<String>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub publishing_status: Option<String>,
    pub etag: Option<String>,
    pub layer_type: Option<String>,
    pub published_access_list: Option<String>,
}

A Layer combines multiple datasources, with styling information, for presentation on a map.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete layers (none)

get layers (response)

cancel processing layers (none)

publish layers (none)

permissions batch delete layers (none)

list published layers (none)

create layers (request|response)

patch layers (request)

permissions batch update layers (none)

permissions list layers (none)

list layers (none)

process layers (none)

parents list layers (none)

get published layers (none)

unpublish layers (none)

Fields

datasources

An array of datasources used to build this layer. If layerType is "image", or layerType is not specified and datasourceType is "image", then each element in this array is a reference to an Image or RasterCollection. If layerType is "vector", or layerType is not specified and datasourceType is "table" then each element in this array is a reference to a Vector Table.

tags

Tags of this Layer.

draft_access_list

Deprecated: The name of an access list of the Map Editor type. The user on whose behalf the request is being sent must be an editor on that access list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.

datasource_type

Deprecated: The type of the datasources used to build this Layer. Note: This has been replaced by layerType, but is still available for now to maintain backward compatibility.

writers_can_edit_permissions

If true, WRITERs of the asset are able to edit the asset permissions.

creator_email

The email address of the creator of this layer. This is only returned on GET requests and not LIST requests.

bbox

A rectangular bounding box which contains all of the data in this Layer. The box is expressed as "west, south, east, north". The numbers represent latitude and longitude in decimal degrees.

last_modifier_email

The email address of the last modifier of this layer. This is only returned on GET requests and not LIST requests.

last_modified_time

The last modified time of this layer. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this Layer.

description

The description of this Layer, supplied by the author.

style

The styling information for a vector layer. Note: Style information is returned in response to a get request but not a list request. After requesting a list of layers, you'll need to send a get request to retrieve the VectorStyles for each layer.

name

The name of this Layer, supplied by the author.

processing_status

The processing status of this layer.

project_id

The ID of the project that this Layer is in.

creation_time

The creation time of this layer. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

publishing_status

The publishing status of this layer.

etag

The ETag, used to refer to the current version of the asset.

layer_type

The type of the datasources used to build this Layer. This should be used instead of datasourceType. At least one of layerType and datasourceType and must be specified, but layerType takes precedence.

published_access_list

Deprecated: The access list to whom view permissions are granted. The value must be the name of a Maps Engine access list of the Map Viewer type, and the user must be a viewer on that list. Note: Google Maps Engine no longer uses access lists. Instead, each asset has its own list of permissions. For backward compatibility, the API still accepts access lists for projects that are already using access lists. If you created a GME account/project after July 14th, 2014, you will not be able to send API requests that include access lists. Note: This is an input field only. It is not returned in response to a list or get request.