Struct google_mapsengine1::LayerMethods

pub struct LayerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on layer resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `cancel_processing(...)`, `create(...)`, `delete(...)`, `get(...)`, `get_published(...)`, `list(...)`, `list_published(...)`, `parents_list(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `process(...)`, `publish(...)` and `unpublish(...)`
// to build up your call.
let rb = hub.layers();

Methods

impl<'a, C, A> LayerMethods<'a, C, A>

fn unpublish(&self, id: &str) -> LayerUnpublishCall<'a, C, A>

Create a builder to help you perform the following task:

Unpublish a layer asset.

Arguments

id - The ID of the layer.

fn publish(&self, id: &str) -> LayerPublishCall<'a, C, A>

Create a builder to help you perform the following task:

Publish a layer asset.

Arguments

id - The ID of the layer.

fn list_published(&self) -> LayerListPublishedCall<'a, C, A>

Create a builder to help you perform the following task:

Return all published layers readable by the current user.

fn create(&self, request: Layer) -> LayerCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create a layer asset.

Arguments

request - No description provided.

fn get_published(&self, id: &str) -> LayerGetPublishedCall<'a, C, A>

Create a builder to help you perform the following task:

Return the published metadata for a particular layer.

Arguments

id - The ID of the layer.

fn delete(&self, id: &str) -> LayerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a layer.

Arguments

id - The ID of the layer. Only the layer creator or project owner are permitted to delete. If the layer is published, or included in a map, the request will fail. Unpublish the layer, and remove it from all maps prior to deleting.

fn get(&self, id: &str) -> LayerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return metadata for a particular layer.

Arguments

id - The ID of the layer.

fn permissions_batch_delete(&self, request: PermissionsBatchDeleteRequest, id: &str) -> LayerPermissionBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

Arguments

request - No description provided.

id - The ID of the asset from which permissions will be removed.

fn permissions_batch_update(&self, request: PermissionsBatchUpdateRequest, id: &str) -> LayerPermissionBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the asset to which permissions will be added.

fn list(&self) -> LayerListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all layers readable by the current user.

fn parents_list(&self, id: &str) -> LayerParentListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all parent ids of the specified layer.

Arguments

id - The ID of the layer whose parents will be listed.

fn cancel_processing(&self, id: &str) -> LayerCancelProcessingCall<'a, C, A>

Create a builder to help you perform the following task:

Cancel processing on a layer asset.

Arguments

id - The ID of the layer.

fn patch(&self, request: Layer, id: &str) -> LayerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Mutate a layer asset.

Arguments

request - No description provided.

id - The ID of the layer.

fn permissions_list(&self, id: &str) -> LayerPermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

Arguments

id - The ID of the asset whose permissions will be listed.

fn process(&self, id: &str) -> LayerProcesCall<'a, C, A>

Create a builder to help you perform the following task:

Process a layer asset.

Arguments

id - The ID of the layer.