Struct google_mapsengine1::MapMethods

pub struct MapMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on map resources. It is not used directly, but through the MapsEngine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mapsengine1 as mapsengine1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mapsengine1::MapsEngine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = MapsEngine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `get_published(...)`, `list(...)`, `list_published(...)`, `patch(...)`, `permissions_batch_delete(...)`, `permissions_batch_update(...)`, `permissions_list(...)`, `publish(...)` and `unpublish(...)`
// to build up your call.
let rb = hub.maps();

Methods

impl<'a, C, A> MapMethods<'a, C, A>

fn list(&self) -> MapListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all maps readable by the current user.

fn list_published(&self) -> MapListPublishedCall<'a, C, A>

Create a builder to help you perform the following task:

Return all published maps readable by the current user.

fn patch(&self, request: Map, id: &str) -> MapPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Mutate a map asset.

Arguments

request - No description provided.

id - The ID of the map.

fn permissions_batch_update(&self, request: PermissionsBatchUpdateRequest, id: &str) -> MapPermissionBatchUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Add or update permission entries to an already existing asset.

An asset can hold up to 20 different permission entries. Each batchInsert request is atomic.

Arguments

request - No description provided.

id - The ID of the asset to which permissions will be added.

fn delete(&self, id: &str) -> MapDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Delete a map.

Arguments

id - The ID of the map. Only the map creator or project owner are permitted to delete. If the map is published the request will fail. Unpublish the map prior to deleting.

fn unpublish(&self, id: &str) -> MapUnpublishCall<'a, C, A>

Create a builder to help you perform the following task:

Unpublish a map asset.

Arguments

id - The ID of the map.

fn publish(&self, id: &str) -> MapPublishCall<'a, C, A>

Create a builder to help you perform the following task:

Publish a map asset.

Arguments

id - The ID of the map.

fn permissions_list(&self, id: &str) -> MapPermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Return all of the permissions for the specified asset.

Arguments

id - The ID of the asset whose permissions will be listed.

fn create(&self, request: Map) -> MapCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Create a map asset.

Arguments

request - No description provided.

fn get_published(&self, id: &str) -> MapGetPublishedCall<'a, C, A>

Create a builder to help you perform the following task:

Return the published metadata for a particular map.

Arguments

id - The ID of the map.

fn get(&self, id: &str) -> MapGetCall<'a, C, A>

Create a builder to help you perform the following task:

Return metadata for a particular map.

Arguments

id - The ID of the map.

fn permissions_batch_delete(&self, request: PermissionsBatchDeleteRequest, id: &str) -> MapPermissionBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Remove permission entries from an already existing asset.

Arguments

request - No description provided.

id - The ID of the asset from which permissions will be removed.