Struct google_mapsengine1::RasterCollectionsRaster

pub struct RasterCollectionsRaster {
    pub description: Option<String>,
    pub tags: Option<Vec<String>>,
    pub project_id: Option<String>,
    pub creation_time: Option<String>,
    pub name: Option<String>,
    pub bbox: Option<Vec<f64>>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub raster_type: Option<String>,
}

A raster resource.

This type is not used in any activity, and only used as part of another schema.

Fields

description

The description of this Raster, supplied by the author.

tags

Tags of this Raster.

project_id

The ID of the project that this Raster is in.

creation_time

The creation time of this raster. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

name

The name of this Raster, supplied by the author.

bbox

A rectangular bounding box which contains all of the data in this Raster. The box is expressed as \"west, south, east, north\". The numbers represent latitudes and longitudes in decimal degrees.

last_modified_time

The last modified time of this raster. The value is an RFC 3339 formatted date-time value (e.g. 1970-01-01T00:00:00Z).

id

A globally unique ID, used to refer to this Raster.

raster_type

The type of this Raster. Always "image" today.